fix:mv test case and handle edge case
